@@ -17,7 +17,7 @@
     "react-autobind": "^1.0.6",
     "react-dom": "^0.14.6",
     "superagent": "^3.5.2",
-    "webpack": "^1.12.9",
+    "webpack": "^2.2.1",
     "webpack-dev-server": "^1.14.1"
   },
   "devDependencies": {
@@ -29,6 +29,7 @@
   },
   "scripts": {
     "dev": "webpack-dev-server --content-base src --inline --hot",
+    "build": "webpack",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
